create branch api

+++ src/main/java/com/mycompany/myapp/respone/branch/BranchCountResponse.java
@@ -2,19 +2,19 @@ package com.mycompany.myapp.respone.branch;
 
 public class BranchCountResponse {
 
-    private Integer count;
+    private Long count;
 
     public BranchCountResponse() {}
 
-    public BranchCountResponse(Integer count) {
+    public BranchCountResponse(Long count) {
         this.count = count;
     }
 
-    public Integer getCount() {
+    public Long getCount() {
         return count;
     }
 
-    public void setCount(Integer count) {
+    public void setCount(Long count) {
         this.count = count;
     }
 }
